Rename rxtx_loop() to rxtx_ring_loop() and migrate appropriately.
