+++ rxtx.c
@@ -12,42 +12,31 @@
 
 #include "rxtx.h"
 
-#include "rxtx_error.h"    // for RXTX_ERRBUF_SIZE, RXTX_ERROR
-#include "rxtx_ring.h" // for rxtx_ring_clear_unreliable_packets_in_buffer(),
-                       //     rxtx_ring_destroy(), rxtx_ring_init(),
+#include "rxtx_error.h" // for RXTX_ERRBUF_SIZE, RXTX_ERROR
+#include "rxtx_ring.h" // for rxtx_ring_destroy(), rxtx_ring_init(),
                        //     rxtx_ring_mark_packets_in_buffer_as_unreliable(),
                        //     rxtx_ring_savefile_open()
-#include "rxtx_savefile.h" // for rxtx_savefile_dump()
-#include "rxtx_stats.h"    // for rxtx_stats_destroy_with_mutex(),
-                           //     rxtx_stats_get_packets_received(),
-                           //     rxtx_stats_get_packets_unreliable(),
-                           //     rxtx_stats_increment_packets_received(),
-                           //     rxtx_stats_increment_packets_unreliable(),
-                           //     rxtx_stats_init_with_mutex()
+#include "rxtx_stats.h" // for rxtx_stats_destroy_with_mutex(),
+                        //     rxtx_stats_get_packets_received(),
+                        //     rxtx_stats_get_packets_unreliable(),
+                        //     rxtx_stats_increment_packets_received(),
+                        //     rxtx_stats_increment_packets_unreliable(),
+                        //     rxtx_stats_init_with_mutex()
 
 #include "interface.h" // for interface_set_promisc_on()
 #include "sig.h"       // for keep_running
 
-#include <linux/if_packet.h> // for PACKET_OUTGOING, sockaddr_ll
-#include <net/if.h>          // for if_nametoindex()
-#include <sys/socket.h>      // for recvfrom(), setsockopt(), sockaddr,
+#include <net/if.h>     // for if_nametoindex()
+#include <sys/socket.h> // for setsockopt()
 
-#include <pcap.h>     // for bpf_u_int32, PCAP_D_IN, PCAP_D_OUT, pcap_pkthdr
-#include <pthread.h>  // for pthread_self()
-#include <sched.h>    // for sched_getcpu()
-#include <stdio.h>    // for fprintf(), NULL, stderr
-#include <stdlib.h>   // for calloc(), exit(), free()
-#include <time.h>     // for time()
-#include <unistd.h>   // for getpid()
+#include <sched.h>  // for sched_getcpu()
+#include <stdio.h>  // for fprintf(), NULL, stderr
+#include <stdlib.h> // for calloc(), exit(), free()
+#include <unistd.h> // for getpid()
 
 #define EXIT_FAIL 1
 
 #define INCREMENT_STEP 1
-
-#define PACKET_BUFFER_SIZE 65535
-
-#define packet_direction_is_rx(sll) (!packet_direction_is_tx(sll))
-#define packet_direction_is_tx(sll) ((sll)->sll_pkttype == PACKET_OUTGOING)
 
 char errbuf[RXTX_ERRBUF_SIZE] = "";
 char *program_basename = NULL;
@@ -299,65 +288,3 @@ void rxtx_set_breakloop_global(void) {
   rxtx_breakloop = 1;
 }
 /* ----------------------------- end of setters ---------------------------- */
-
-/* ========================================================================= */
-void *rxtx_loop(void *ring) {
-  struct rxtx_ring *p = ring;
-
-  if (rxtx_verbose_isset(p->rtd)) {
-    fprintf(stderr, "Worker '%lu' handling ring '%d' running on cpu '%d'.\n",
-                                       pthread_self(), p->idx, sched_getcpu());
-  }
-
-  rxtx_ring_clear_unreliable_packets_in_buffer(p);
-
-  while (!rxtx_breakloop_isset(p->rtd)) {
-
-    if (rxtx_packet_count_reached(p->rtd)) {
-      break;
-    }
-
-    struct sockaddr_ll sll;
-    unsigned char packet[PACKET_BUFFER_SIZE];
-
-    unsigned int sll_length = sizeof(sll);
-    int packet_length = recvfrom(p->fd, packet, sizeof(packet), 0,
-                                         (struct sockaddr *)&sll, &sll_length);
-
-    if (packet_length == -1) {
-      continue;
-    }
-
-    if (rxtx_get_direction(p->rtd) == PCAP_D_OUT &&
-                                                packet_direction_is_rx(&sll)) {
-      continue;
-    }
-
-    if (rxtx_get_direction(p->rtd) == PCAP_D_IN &&
-                                                packet_direction_is_tx(&sll)) {
-      continue;
-    }
-
-    rxtx_increment_packets_received(p->rtd);
-    rxtx_stats_increment_packets_received(p->stats, INCREMENT_STEP);
-
-    if (p->savefile) {
-      /* no need for memset(), we're initializing every member */
-      struct pcap_pkthdr header;
-      header.caplen     = (bpf_u_int32)packet_length;
-      header.len        = (bpf_u_int32)packet_length;
-      header.ts.tv_sec  = time(NULL);
-      header.ts.tv_usec = 0;
-
-      int status;
-      status = rxtx_savefile_dump(p->savefile, &header, packet,
-                                           rxtx_packet_buffered_isset(p->rtd));
-
-      if (status == RXTX_ERROR) {
-        fprintf(stderr, "%s: %s\n", program_basename, p->errbuf);
-        exit(EXIT_FAIL);
-      }
-    }
-  }
-  return NULL;
-}
